+++ packages/web/src/javascripts/Components/SuperEditor/Plugins/RemoteImagePlugin/RemoteImageComponent.tsx
@@ -4,13 +4,16 @@ import Spinner from '@/Components/Spinner/Spinner'
 import { isDesktopApplication } from '@/Utils'
 import { BlockWithAlignableContents } from '@lexical/react/LexicalBlockWithAlignableContents'
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
-import { classNames, Platform } from '@standardnotes/snjs'
+import { classNames, Platform, PrefKey } from '@standardnotes/snjs'
 import { $getNodeByKey, CLICK_COMMAND, COMMAND_PRIORITY_LOW, ElementFormatType, NodeKey } from 'lexical'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { $createFileNode } from '../EncryptedFilePlugin/Nodes/FileUtils'
 import { RemoteImageNode } from './RemoteImageNode'
 import { isIOS } from '@standardnotes/ui-services'
 import { useLexicalNodeSelection } from '@lexical/react/useLexicalNodeSelection'
+import usePreference from '@/Hooks/usePreference'
+import { getCSSValueFromAlignment, ImageAlignmentOptions } from '@/Components/FilePreview/ImageAlignmentOptions'
+import { getOverflows } from '@/Components/Popover/Utils/Collisions'
 
 type Props = {
   src: string
@@ -21,10 +24,11 @@ type Props = {
     focus: string
   }>
   format: ElementFormatType | null
+  setFormat: (format: ElementFormatType) => void
   nodeKey: NodeKey
 }
 
-const RemoteImageComponent = ({ className, src, alt, node, format, nodeKey }: Props) => {
+const RemoteImageComponent = ({ className, src, alt, node, format, nodeKey, setFormat }: Props) => {
   const application = useApplication()
   const [editor] = useLexicalComposerContext()
 
@@ -91,9 +95,30 @@ const RemoteImageComponent = ({ className, src, alt, node, format, nodeKey }: Pr
     )
   }, [editor, isSelected, nodeKey, setSelected])
 
+  const changeAlignment = useCallback(
+    (format: ElementFormatType) => {
+      editor.update(() => {
+        setFormat(format)
+      })
+    },
+    [editor, setFormat],
+  )
+
+  const defaultSuperImageAlignment = usePreference(PrefKey.SuperNoteImageAlignment)
+  const finalAlignment = format || defaultSuperImageAlignment
+  const alignItems: 'start' | 'center' | 'end' = getCSSValueFromAlignment(finalAlignment)
+
   return (
     <BlockWithAlignableContents className={className} format={format} nodeKey={nodeKey}>
-      <div ref={ref} className="relative flex min-h-[2rem] flex-col items-center gap-2.5">
+      <div
+        ref={ref}
+        className="group relative flex min-h-[2rem] flex-col gap-2.5"
+        style={{ alignItems }}
+        onClick={(e) => {
+          e.preventDefault()
+          e.stopPropagation()
+        }}
+      >
         <img
           alt={alt}
           src={src}
@@ -101,6 +126,25 @@ const RemoteImageComponent = ({ className, src, alt, node, format, nodeKey }: Pr
             setDidImageLoad(true)
           }}
         />
+        <div
+          className="invisible absolute bottom-full left-1/2 z-10 w-max -translate-x-1/2 px-1 pb-1 focus-within:visible group-hover:visible [.embedBlockFocused_&]:visible"
+          ref={(popover) => {
+            const editorRoot = editor.getRootElement()
+            if (!popover || !editorRoot) {
+              return
+            }
+            const editorRootRect = editorRoot.getBoundingClientRect()
+            const popoverRect = popover.getBoundingClientRect()
+            const overflows = getOverflows(popoverRect, editorRootRect)
+            if (overflows.top > 0) {
+              popover.style.setProperty('--tw-translate-y', `${overflows.top}px`)
+            }
+          }}
+        >
+          <div className="flex gap-1 rounded border border-border bg-default px-1 py-0.5">
+            <ImageAlignmentOptions alignment={finalAlignment} changeAlignment={changeAlignment} />
+          </div>
+        </div>
         {didImageLoad && canShowSaveButton && (
           <button
             className={classNames(