+++ packages/clipper/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.523](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.522...@standardnotes/clipper@1.1.523) (2025-06-18)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.522](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.521...@standardnotes/clipper@1.1.522) (2025-06-16)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.521](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.520...@standardnotes/clipper@1.1.521) (2025-06-12)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.520](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.519...@standardnotes/clipper@1.1.520) (2025-06-05)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.519](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.518...@standardnotes/clipper@1.1.519) (2025-05-22)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
 ## [1.1.518](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.517...@standardnotes/clipper@1.1.518) (2025-05-12)
 
 **Note:** Version bump only for package @standardnotes/clipper