+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.110.128](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.197.0...@standardnotes/desktop@3.110.128) (2025-06-18)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.127](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.196.9...@standardnotes/desktop@3.110.127) (2025-06-16)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.126](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.196.8...@standardnotes/desktop@3.110.126) (2025-06-12)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.125](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.196.7...@standardnotes/desktop@3.110.125) (2025-06-05)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.124](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.196.6...@standardnotes/desktop@3.110.124) (2025-05-22)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.110.123](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.196.5...@standardnotes/desktop@3.110.123) (2025-05-12)
 
 **Note:** Version bump only for package @standardnotes/desktop