+++ packages/files/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.20.12](https://github.com/standardnotes/app/compare/@standardnotes/files@1.20.11...@standardnotes/files@1.20.12) (2025-06-18)
+
+**Note:** Version bump only for package @standardnotes/files
+
+## [1.20.11](https://github.com/standardnotes/app/compare/@standardnotes/files@1.20.10...@standardnotes/files@1.20.11) (2025-06-16)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.20.10](https://github.com/standardnotes/app/compare/@standardnotes/files@1.20.9...@standardnotes/files@1.20.10) (2024-05-07)
 
 **Note:** Version bump only for package @standardnotes/files